feat: roles and auth router
